refactor: mv with storage sync
